pyproject.toml
[project]
name = "esp-miao"
version = "0.5.1"
description = "ESP32 Voice Agent with esp-sr + Ollama LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "matplotlib>=3.10.8",
    "ollama>=0.6.1",
    "paho-mqtt>=2.1.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]

[project.scripts]
esp-miao = "esp_miao.server:main"
esp32-sim = "esp_miao.esp32_simulator:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
